Refactored stmgr and instance related communication to seperate classes
