+++ heron/stmgr/src/cpp/BUILD
@@ -72,6 +72,7 @@ cc_library(
         "manager/stmgr-client.cpp",
         "manager/stmgr-clientmgr.cpp",
         "manager/stmgr-server.cpp",
+        "manager/instance-server.cpp",
         "manager/stmgr.cpp",
         "manager/stream-consumers.cpp",
         "manager/tmaster-client.cpp",
@@ -82,6 +83,7 @@ cc_library(
         "manager/stmgr-client.h",
         "manager/stmgr-clientmgr.h",
         "manager/stmgr-server.h",
+        "manager/instance-server.h",
         "manager/stmgr.h",
         "manager/stream-consumers.h",
         "manager/tmaster-client.h",
