+++ heron/stmgr/src/cpp/manager/tmaster-client.h
@@ -30,7 +30,8 @@ namespace stmgr {
 class TMasterClient : public Client {
  public:
   TMasterClient(EventLoop* eventLoop, const NetworkOptions& _options, const sp_string& _stmgr_id,
-                const sp_string& _stmgr_host, sp_int32 _stmgr_port, sp_int32 _shell_port,
+                const sp_string& _stmgr_host, sp_int32 _data_port, sp_int32 _local_data_port,
+                sp_int32 _shell_port,
                 VCallback<proto::system::PhysicalPlan*> _pplan_watch,
                 VCallback<sp_string> _stateful_checkpoint_watch,
                 VCallback<sp_string, sp_int64> _restore_topology_watch,
@@ -82,7 +83,8 @@ class TMasterClient : public Client {
 
   sp_string stmgr_id_;
   sp_string stmgr_host_;
-  sp_int32 stmgr_port_;
+  sp_int32 data_port_;
+  sp_int32 local_data_port_;
   sp_int32 shell_port_;
 
   bool register_request_set_;
